+++ src/SqlArtisan/Internal/SqlPart/Clause/GroupBy/GroupingElement.cs
@@ -8,4 +8,7 @@ namespace SqlArtisan.Internal;
 /// </summary>
 public abstract class GroupingElement : SqlPart
 {
+    private protected GroupingElement()
+    {
+    }
 }