+++ src/SqlArtisan/Internal/SqlPart/Condition/ArrayCondition/ArrayCondition.cs
@@ -1,13 +1,17 @@
 namespace SqlArtisan.Internal;
 
-public abstract class ArrayCondition(
-    SqlExpression leftSide,
-    string @operator,
-    SqlExpression rightSide) : SqlCondition
+public abstract class ArrayCondition : SqlCondition
 {
-    private readonly SqlExpression _leftSide = leftSide;
-    private readonly string _operator = @operator;
-    private readonly SqlExpression _rightSide = rightSide;
+    private readonly SqlExpression _leftSide;
+    private readonly string _operator;
+    private readonly SqlExpression _rightSide;
+
+    private protected ArrayCondition(SqlExpression leftSide, string @operator, SqlExpression rightSide)
+    {
+        _leftSide = leftSide;
+        _operator = @operator;
+        _rightSide = rightSide;
+    }
 
     internal override void Format(SqlBuildingBuffer buffer) => buffer
         .Append(_leftSide)