+++ src/SqlArtisan/Internal/SqlPart/Expression/BinaryOperator.cs
@@ -1,13 +1,17 @@
 namespace SqlArtisan.Internal;
 
-public abstract class BinaryOperator(
-    SqlExpression leftSide,
-    string @operator,
-    SqlExpression rightSide) : SqlExpression
+public abstract class BinaryOperator : SqlExpression
 {
-    private readonly SqlExpression _leftSide = leftSide;
-    private readonly string _operator = @operator;
-    private readonly SqlExpression _rightSide = rightSide;
+    private readonly SqlExpression _leftSide;
+    private readonly string _operator;
+    private readonly SqlExpression _rightSide;
+
+    private protected BinaryOperator(SqlExpression leftSide, string @operator, SqlExpression rightSide)
+    {
+        _leftSide = leftSide;
+        _operator = @operator;
+        _rightSide = rightSide;
+    }
 
     internal override void Format(SqlBuildingBuffer buffer) => buffer
         .OpenParenthesis(_leftSide)